Initial Commit of Phoenix Project
I am working on this project to primarily learn elixer. But it will be
nice to learn a new framework while i'm at it. And also keep my
React/Redux skills sharp.

All of these files add the intial setup of the phoenix project (in
'server' directory) and the React client side implementation (in
'client' directory).