Adding React Router
'index.js' now contains a 'Router' component, and setups up the routes
to work with 'App', 'Home', and 'Settings'.

The 'Route' component of 'react-router' requires a component. This
component, 'App', is essentially just a container that renders
components that it receives from the router (in the props). When you
access each route, the component of that route is passed into 'App' as a
child, which 'App' then just renders.